fastapi/requirements.txt
aenum==3.1.15
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.10.4
arrow==1.3.0
attrs==23.2.0
beanie==1.25.0
boto3==1.37.36
botocore==1.37.36
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.14
dictdiffer==0.9.0
dnspython==2.6.1
dotenv==0.9.9
fastapi==0.115.12
fastapi_basic==0.1.3
frozenlist==1.4.1
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.26.0
idna==3.10
Jinja2==3.1.5
jmespath==1.0.1
lazy-model==0.2.0
line-bot-sdk==3.12.0
MarkupSafe==3.0.2
marshmallow==3.21.3
motor==3.4.0
multidict==6.0.5
packaging==24.1
pydantic==2.11.3
pydantic-settings==2.2.1
pydantic_core==2.33.1
pymongo==4.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
s3transfer==0.11.5
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
toml==0.10.2
types-python-dateutil==2.9.0.20240316
typing-inspection==0.4.0
typing_extensions==4.13.1
tzlocal==5.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.19.0
watchfiles==0.21.0
watchtower==3.4.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
